helper: Remove unnecessary cast

+++ src/helpers.rs
@@ -1,7 +1,7 @@
 use std::ffi::CString;
 
 pub fn c_str_with_size(size: usize) -> CString {
-    let mut buffer = Vec::with_capacity(size as usize + 1);
-    buffer.extend([b' '].iter().cycle().take(size as usize));
+    let mut buffer = Vec::with_capacity(size + 1);
+    buffer.extend([b' '].iter().cycle().take(size));
     return unsafe { CString::from_vec_unchecked(buffer) };
 }
